Fechas funcionando con controlador
